+++ tests/conftest.py
@@ -6,7 +6,5 @@ from dotenv import load_dotenv
 
 @pytest.fixture(scope="session", autouse=True)
 def load_env():
-    dotenv_path = (
-        Path(__file__).parent / "../.env"
-    )  # Construct path from current test file dir
+    dotenv_path = Path(__file__).parent / "../.env"  # Construct path from current test file dir
     load_dotenv(dotenv_path=dotenv_path)